feat: add authentication and speech features

+++ api/delete.php
@@ -1,8 +1,10 @@
 <?php
+session_start();
 
 $rawData = file_get_contents("php://input");
 $data = json_decode($rawData, true);
 $diaryId = $data["currentDiaryId"];
+$userId = $_SESSION["user_id"];
 
 //データベース接続
 $dsn = "mysql:dbname=linguaDiary;host=localhost;charset=utf8mb4";
@@ -17,12 +19,13 @@ try {
 
 //削除処理
 $stmt = $pdo->prepare (
-  "DELETE FROM diaries WHERE id = :id"
+  "DELETE FROM diaries WHERE id = :id AND user_id = :userId"
 );
 
-$stmt->execute (
-  [":id" => $diaryId]
-);
+$stmt->execute ([
+  ":id" => $diaryId,
+  ":user_id" => $userId
+]);
 
 $deletedCount = $stmt->rowCount();
 if ($deletedCount > 0) {
